pyproject.toml
[tool.poetry]
name = "duckcli"
version = "0.0.4"
description = "API-driven rich network and system CLI tool"
authors = ["Prathap Shetty"]
license = "MIT"
readme = "README.md"

packages = [
    { include = "duckcli", from = "src" }
]

maintainers = [
    "Prathap Shetty <prathap.shetty@duckcli.com>"
   
]
homepage = "https://duckcli.com/"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.85.0"
rich = "12.6.0"
click = "8.1.3"
typer = "0.6.1"
requests = "2.28.1"
pydantic =  "1.10.2"
urllib3 = "1.26.12"
nested-lookup = "0.2.25"
GitPython = "3.1.18"
SQLAlchemy = "1.4.41"
pyats = "22.9"
genie = "22.9"
pyats-genie-command-parse = "1.3.2"
ncclient = "0.6.13"
netaddr = "0.8.0"
paramiko = "2.11.0"
netmiko = "4.1.2"
uvicorn = "0.18.3"
validators = "0.20.0"
python-jose = "3.3.0"
passlib = "1.7.4"
python-multipart = "0.0.5"
icmplib = "3.0.3"
email-validator = "1.3.0"

[tool.poetry.dev-dependencies]
schemathesis = "3.17.3"
hypothesis = "6.56.3"
hypothesis-graphql = "0.9.1"
hypothesis-jsonschema = "0.22.0"
pytest = "7.1.3"
pytest-cov = "4.0.0"
pytest-subtests = "0.7.0"

[tool.poetry.scripts]
duck-cli = "duckcli.frontend.main:main"
duck-cli-admin = "duckcli.backend.core.utils.admin:main"


[tool.pytest.ini_options]
addopts = "--cov=duckcli"
testpaths = [
    "tests",
]

[tool.mypy]
mypy_path = "src"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
